packages/semantic-normalization/pyproject.toml
[project]
name = "semantic-normalization"
version = "1.0.0"
description = "LLM-based semantic normalization layer for speech-to-act system"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0"
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
